+++ packages/nx-plugin/src/generators/init/generator.int.test.ts
@@ -1,6 +1,5 @@
 import { type Tree, logger, readJson, readNxJson } from '@nx/devkit';
 import { createTreeWithEmptyWorkspace } from '@nx/devkit/testing';
-import { describe, expect, it, vi } from 'vitest';
 import { initGenerator } from './generator.js';
 
 type PackageJson = {